requirements.txt
# Ultralytics requirements
# Usage: pip install -r requirements.txt

# Base ----------------------------------------
numpy==1.26.0
matplotlib>=3.2.2
opencv-python>=4.6.0
Pillow>=7.1.2
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
tqdm>=4.64.0

# Plotting ------------------------------------
pandas>=1.1.4
seaborn>=0.11.0

# Extras --------------------------------------
psutil
pyqt5==5.12.3
labelme
labelimg
matplotlib
opencv-contrib-python
opencv-python
tensorboard
dill
ultralytics==8.3.28
einops==0.8.0
pycocotools==2.0.8